Add Netlify badge and improve code readability
Added Netlify badge to the footer of the left navigation bar across multiple pages as a mark of acknowledgement for the service. This was achieved by creating a new `LeftNavFooter` component and importing it in necessary files. Additionally, minor indentation and syntax adjustments have been made in these files for improved code readability and overall code quality.