Update Netlify logo in Left Nav Footer
The Netlify logo in the LeftNavFooter component has been updated to netlify-light.svg instead of netlify-color-bg.svg. This change makes the logo more visually consistent with the rest of the application's design. The new logo is clearer and easier on the eyes when viewed against the background color of the application's left navigation bar.